+++ Sources/Auth/Internal/SessionManager.swift
@@ -5,41 +5,16 @@ struct SessionRefresher: Sendable {
   var refreshSession: @Sendable (_ refreshToken: String) async throws -> Session
 }
 
-struct SessionManager: Sendable {
-  var session: @Sendable (_ shouldValidateExpiration: Bool) async throws -> Session
-  var update: @Sendable (_ session: Session) async throws -> Void
-  var remove: @Sendable () async -> Void
-}
-
-extension SessionManager {
-  func session(shouldValidateExpiration: Bool = true) async throws -> Session {
-    try await session(shouldValidateExpiration)
-  }
-}
-
-extension SessionManager {
-  static let live: SessionManager = {
-    let manager = _DefaultSessionManager()
-
-    return SessionManager(
-      session: { try await manager.session(shouldValidateExpiration: $0) },
-      update: { try await manager.update($0) },
-      remove: { await manager.remove() }
-    )
-  }()
-}
-
-private actor _DefaultSessionManager {
+actor SessionManager {
   private var task: Task<Session, any Error>?
 
-  @Dependency(\.sessionStorage)
-  private var storage: SessionStorage
+  private var storage: any AuthLocalStorage {
+    Current.configuration.localStorage
+  }
 
-  @Dependency(\.sessionRefresher)
-  private var sessionRefresher: SessionRefresher
-
-  @Dependency(\.logger)
-  private var logger: (any SupabaseLogger)?
+  private var sessionRefresher: SessionRefresher {
+    Current.sessionRefresher
+  }
 
   func session(shouldValidateExpiration: Bool) async throws -> Session {
 
@@ -55,7 +30,6 @@ private actor _DefaultSessionManager {
     if let task {
       return try await task.value
     }
-
     task = Task {
       defer { task = nil }