+++ Sources/Functions/Types.swift
@@ -1,3 +1,4 @@
+import _Helpers
 import Foundation
 
 /// An error type representing various errors that can occur while invoking functions.
@@ -23,7 +24,7 @@ public struct FunctionInvokeOptions: Sendable {
   /// Method to use in the function invocation.
   let method: Method?
   /// Headers to be included in the function invocation.
-  let headers: [String: String]
+  let headers: HTTPHeaders
   /// Body data to be sent with the function invocation.
   let body: Data?
   /// The Region to invoke the function in.
@@ -43,7 +44,7 @@ public struct FunctionInvokeOptions: Sendable {
     region: String? = nil,
     body: some Encodable
   ) {
-    var defaultHeaders = headers
+    var defaultHeaders = HTTPHeaders()
 
     switch body {
     case let string as String:
@@ -59,7 +60,7 @@ public struct FunctionInvokeOptions: Sendable {
     }
 
     self.method = method
-    self.headers = defaultHeaders.merging(headers) { _, new in new }
+    self.headers = defaultHeaders.merged(with: HTTPHeaders(headers))
     self.region = region
   }
 
@@ -76,7 +77,7 @@ public struct FunctionInvokeOptions: Sendable {
     region: String? = nil
   ) {
     self.method = method
-    self.headers = headers
+    self.headers = HTTPHeaders(headers)
     self.region = region
     body = nil
   }
@@ -87,6 +88,23 @@ public struct FunctionInvokeOptions: Sendable {
     case put = "PUT"
     case patch = "PATCH"
     case delete = "DELETE"
+  }
+
+  var httpMethod: HTTPMethod? {
+    switch method {
+    case .get:
+      .get
+    case .post:
+      .post
+    case .put:
+      .put
+    case .patch:
+      .patch
+    case .delete:
+      .delete
+    case nil:
+      nil
+    }
   }
 }